apis/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.61
botocore==1.31.61
exceptiongroup==1.1.3
fastapi==0.103.2
idna==3.4
jmespath==1.0.1
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
urllib3==1.26.17
uvicorn
